Change spatial structs to use float, Add AtX, AtY methods to ISegment2D

+++ src/protocol/Edelstein.Protocol.Util.Spatial/IRectangle2D.cs
@@ -5,14 +5,14 @@ public interface IRectangle2D : IObject2D
     IPoint2D P1 { get; }
     IPoint2D P2 { get; }
     
-    int Left { get; }
-    int Right { get; }
-    int Top { get; }
-    int Bottom { get; }
+    float Left { get; }
+    float Right { get; }
+    float Top { get; }
+    float Bottom { get; }
 
-    int Height { get; }
-    int Width { get; }
-    int Area { get; }
+    float Height { get; }
+    float Width { get; }
+    float Area { get; }
     
     bool IsSquare { get; }
